+++ deploy/follow_the_drow/include/main.hpp
@@ -24,7 +24,7 @@ extern std::string DROW_DETECTOR_TOPIC;
 void loadArgumentsForNode(int argc, char** argv, const std::string& name);
 
 // Live loader node arguments
-extern std::map<std::string, double> TRANSFORM_DATA;
+extern std::map<std::string, float> TRANSFORM_DATA;
 extern std::string TOP_LIDAR_TOPIC;
 extern std::string BOTTOM_LIDAR_TOPIC;
 extern std::string ODOMETRY_TOPIC;
